refactor: waitTransaction
