+++ packages/nextjs/lib/utils.ts
@@ -1,7 +1,5 @@
 import { type ClassValue, clsx } from "clsx";
 import { twMerge } from "tailwind-merge";
-import { Schemas } from "~~/services/graphql/config";
-import { pollTransactionQuery } from "~~/services/graphql/fetchers/utils.service";
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
@@ -23,37 +21,4 @@ export function addPrefixToUrl(url: string, prefix: string): string {
     return prefix + url;
   }
   return url;
-}
-
-export async function waitTransaction(schema: Schemas, transactionHash: `0x${string}` | undefined) {
-  if (!transactionHash) {
-    console.warn("No transaction hash provided");
-    return false;
-  }
-
-  const poll = async () => {
-    const baseSleep = 200;
-    const maxSleep = 5000;
-
-    let sleepTime = baseSleep;
-    for (let i = 1; i <= 5; i++) {
-      // Poll the transaction status with schema + "s" to use where clause
-      const result = await pollTransactionQuery(transactionHash, schema + "s");
-      if (result) {
-        return true;
-      }
-      await new Promise(resolve => setTimeout(resolve, sleepTime));
-      sleepTime = Math.min(baseSleep * 2 ** i, maxSleep);
-    }
-
-    console.error("No transaction found with hash:", transactionHash);
-    return false;
-  };
-
-  try {
-    return await poll();
-  } catch (error) {
-    console.error(`Error occurred while waiting for transaction, is the schema name '${schema}' ok??:`, error);
-    return false;
-  }
 }
